Fix tests
Tests were failing on master since I introduced an `order` column on
step which is non nullable but I had not updated the fixture. As a
result all tests were failing. I also removed the test parallelization
since it was masking the true errors and it was unclear why tests were
failing. I will re-add this back when we need it.